Feature/1

+++ src/components/DeleteButton.tsx
@@ -1,8 +1,8 @@
 'use client';
-import { CreateAuthContext } from '@/app/context/AuthContext';
+import { CreateAuthContext } from '@/context/AuthContext';
 import React, { useContext } from 'react';
 import { doc, deleteDoc } from 'firebase/firestore';
-import { cloudDb } from '../../../firebase';
+import { cloudDb } from '../../firebase';
 import { useRouter } from 'next/navigation';
 
 interface Props {
@@ -19,9 +19,12 @@ export default function DeleteButton({ uid, post }: Props) {
       className="hover:font-bold"
       onClick={async () => {
         if (userInformation?.uid === uid) {
-          await deleteDoc(doc(cloudDb, 'posts', `${post}`));
-          router.push(`/blog/all`);
-          router.refresh();
+          const deletePost = confirm('삭제하시겠습니까 ?');
+          if (deletePost === true) {
+            await deleteDoc(doc(cloudDb, 'posts', `${post}`));
+            router.push(`/blog/all`);
+            router.refresh();
+          }
         } else {
           alert('작성자만 삭제 가능합니다.');
         }
